% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\docType{methods}
\name{as.data.frame,SalesData-method}
\alias{as.data.frame,SalesData}
\alias{as.data.frame,SalesData-method}
\title{Coerce Sales Data to Data Frame}
\usage{
\S4method{as.data.frame}{SalesData}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{SalesData object.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
A method to coerce Sales Data to Data Frame.
}
\examples{
x <- new(Class = "SalesData", 
    Date = as.Date(x = 1:3, origin = "2015-12-18"),
    Daily.Total = 1:3,
    Outlet = "CHIPPENHAM")
as.data.frame(x)
}

