% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected.R
\name{expected}
\alias{expected}
\title{Expected Sales}
\usage{
expected(x, when)
}
\arguments{
\item{x}{SalesData object}

\item{when}{Date object specifying Date for which to extrapolate.
If missing, tomorrow's date will be used.}
}
\value{
data frame with columns Date and Daily.Total
}
\description{
Expected Sales using linear extrapolation, 
tomorrow by default.
}
\examples{
x <- new(Class = "SalesData", 
    Date = as.Date(x = 1:3, origin = "2015-12-18"),
    Daily.Total = 1:3,
    Outlet = "CHIPPENHAM")
expected(x)
}

