% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\docType{methods}
\name{print,SalesData-method}
\alias{print,SalesData}
\alias{print,SalesData-method}
\title{Print Sales Data}
\usage{
\S4method{print}{SalesData}(x, ...)
}
\arguments{
\item{x}{SalesData object.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
A method to print an object.
}
\examples{
x <- new(Class = "SalesData", 
    Date = as.Date(x = 1:3, origin = "2015-12-18"),
    Daily.Total = 1:3,
    Outlet = "CHIPPENHAM")
print(x)
}

