% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Data of Motorcycle Sales}
\format{
A tibble with 1000 rows and 8 variables
\describe{
\item{date}{chr Date an item was purchases}
\item{warehouse}{chr Warehouse an item was bought}
\item{client_type}{chr Type of client buying an item}
\item{product_line}{chr Product an item belongs to}
\item{quantity}{dbl Quantity of item bought by a client}
\item{unit_price}{dbl Unit price of an item}
\item{total}{dbl Total price}
\item{payment}{chr Payment Method}
}
}
\usage{
sales
}
\description{
Motorcycle Sales data containing sales of motorcycles parts in various regions
}
\keyword{datasets}
