% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalesData-class.R
\docType{class}
\name{SalesData-class}
\alias{SalesData}
\alias{SalesData-class}
\title{An S4 class to represent Sales Data.}
\description{
An object with three slots.
}
\section{Slots}{

\describe{
\item{\code{Date}}{Date vector}

\item{\code{Daily.Total}}{numeric value of sales of length Date}

\item{\code{Outlet}}{single character name of outlet (e.g. "CHIPPENHAM")}
}}
\examples{
x <- new(Class = "SalesData", 
    Date = as.Date(x = 1:3, origin = "2015-12-18"),
    Daily.Total = 1:3,
    Outlet = "CHIPPENHAM")
print(x)
validObject(object = x)
}

