% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\docType{methods}
\name{unique,SalesData-method}
\alias{unique,SalesData}
\alias{unique,SalesData-method}
\title{Find Unique Sales Data Daily.Total}
\usage{
\S4method{unique}{SalesData}(x)
}
\arguments{
\item{x}{SalesData object.}

\item{\dots}{further arguments passed to or from other methods.}
}
\description{
A method to get unique Daily.Total.
}
\examples{
x <- new(Class = "SalesData", 
    Date = as.Date(x = 1:3, origin = "2015-12-18"),
    Daily.Total = 1:3,
    Outlet = "CHIPPENHAM")
unique(x)
}

